`set_weights_extrinsic` moved to `bittensor/core/extrinsics/commit_weights.py`
